requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
Pillow>=10.4.0
python-decouple==3.8
django-filter==23.3
